+++ src/App.js
@@ -22,6 +22,7 @@ import {
   Stack,
   Modal,
   Lightbox,
+  Drawer,
 } from 'react-native-router-flux';
 import OneSignal from 'react-native-onesignal';
 import codePush, {LocalPackage} from 'react-native-code-push';
@@ -158,8 +159,11 @@ import ModalComboLocation from './components/ModalComboLocation';
 import APIHandler from './network/APIHandler';
 import Transaction from './components/payment/Transaction';
 import ModalFilter from './components/stores/ModalFilter';
+import FilterComponent from './components/stores/FilterDrawer';
 import {News} from './containers/Social';
 import {ModalComment} from './components/Social';
+import SalesReport from './containers/SalesReport/SalesReport';
+import Social from './containers/Social/Social';
 
 /**
  * Not allow font scaling
@@ -728,6 +732,19 @@ class RootRouter extends Component {
         <Overlay key="overlay">
           <Modal key="modal" hideNavBar transitionConfig={getTransitionConfig}>
             <Lightbox key={appConfig.routes.sceneWrapper}>
+              {/* <Drawer
+                key="drawer-filter"
+                hideNavBar={true}
+                drawerPosition="right"
+                drawerWidth={350}
+                contentComponent={FilterComponent}>
+                <Scene
+                  key={`${appConfig.routes.store}_1`}
+                  component={StoreContainer}
+                  {...navBarConfig}
+                  back
+                />
+              </Drawer> */}
               <Scene
                 key="root"
                 titleStyle={{alignSelf: 'center'}}
@@ -828,6 +845,16 @@ class RootRouter extends Component {
                   </Stack>
                 </Tabs>
 
+                {/* ================ SOCIAL ================ */}
+                <Stack key={appConfig.routes.social}>
+                  <Scene
+                    key={`${appConfig.routes.social}_1`}
+                    component={Social}
+                    {...navBarConfig}
+                    back
+                  />
+                </Stack>
+
                 {/* ================ MODAL COMMENT ================ */}
                 <Stack key={appConfig.routes.modalComment}>
                   <Scene
@@ -865,6 +892,15 @@ class RootRouter extends Component {
                     {...navBarConfig}
                     title={t('screen.commissionIncomeStatement.mainTitle')}
                     component={CommissionIncomeStatement}
+                    back
+                  />
+                </Stack>
+                <Stack key={appConfig.routes.salesReport}>
+                  <Scene
+                    key={`${appConfig.routes.salesReport}_1`}
+                    {...navBarConfig}
+                    title={t('screen.salesReport.mainTitle')}
+                    component={SalesReport}
                     back
                   />
                 </Stack>
@@ -1139,14 +1175,20 @@ class RootRouter extends Component {
                   />
                 </Stack>
 
-                <Stack key={appConfig.routes.store}>
+                <Drawer
+                  key={appConfig.routes.store}
+                  hideNavBar={true}
+                  drawerPosition="right"
+                  drawerWidth={350}
+                  hideDrawerButton={true}
+                  contentComponent={(props) => <FilterComponent {...props} />}>
                   <Scene
                     key={`${appConfig.routes.store}_1`}
                     component={StoreContainer}
                     {...navBarConfig}
                     back
                   />
-                </Stack>
+                </Drawer>
 
                 <Stack key="stores_list">
                   <Scene