+++ src/config.js
@@ -12,7 +12,7 @@ class Config {
   }
 
   get tagVersion() {
-    return 'r11.6.2';
+    return 'r11.6.3';
   }
 
   get appName() {
@@ -263,6 +263,9 @@ class Config {
       // Gamification
       lotteryGame: 'lotteryGame',
       filterProduct: 'filterProduct',
+      salesReport: 'saleReport',
+      // Social
+      social: 'social',
     };
   }
 }