+++ src/helper/servicesHandler/types.js
@@ -89,6 +89,7 @@ export const SERVICES_TYPE = {
   COMMISSION_INCOME_STATEMENT: 'commission_income_statement',
 
   /** GAMIFICATION */
-  LOTTERY_GAME: 'lottery_game'
+  LOTTERY_GAME: 'lottery_game',
 
+  SOCIAL: 'social',
 };