+++ src/i18n/translations/index.js
@@ -1,4 +1,4 @@
-import { languages } from '../constants';
+import {languages} from '../constants';
 
 import vi_common from './vi/common.json';
 import en_common from './en/common.json';
@@ -54,6 +54,8 @@ import vi_commissionIncomeStatement from './vi/commissionIncomeStatement.json';
 import en_commissionIncomeStatement from './en/commissionIncomeStatement.json';
 import vi_social from './vi/social.json';
 import en_social from './en/social.json';
+import vi_salesReport from './vi/salesReport.json';
+import en_salesReport from './en/salesReport.json';
 
 export default {
   [languages.vi.value]: {
@@ -83,7 +85,8 @@ export default {
     resetPassword: vi_resetPassword,
     rateApp: vi_rateApp,
     commissionIncomeStatement: vi_commissionIncomeStatement,
-    social: vi_social
+    social: vi_social,
+    salesReport: vi_salesReport,
   },
   [languages.en.value]: {
     common: en_common,
@@ -112,6 +115,7 @@ export default {
     resetPassword: en_resetPassword,
     rateApp: en_rateApp,
     commissionIncomeStatement: en_commissionIncomeStatement,
-    social: en_social
-  }
+    social: en_social,
+    salesReport: en_salesReport,
+  },
 };