+++ src/network/API/CommonAPI.js
@@ -436,7 +436,7 @@ class CommonAPI {
   get FILTER_URL_TAG() {
     return BaseAPI.apiDomain + 'apiSite/get_tags/';
   }
-  
+
   get USER_LIST_NEWS_CATEGORY() {
     return BaseAPI.apiDomain + 'apiUser/list_news_category';
   }
@@ -447,11 +447,25 @@ class CommonAPI {
   }
 
   get SOCIAL_COMMENT() {
-    return BaseAPI.socialDomain + 'comments/store'
+    return BaseAPI.socialDomain + 'comments/store';
   }
 
   get SOCIAL_LIKES() {
-    return BaseAPI.socialDomain + 'likes/store'
+    return BaseAPI.socialDomain + 'likes/store';
+  }
+  get INVITED_REVENUE() {
+    return BaseAPI.apiDomain + 'apiUser/invited_revenue';
+  }
+  get SOCIAL_GROUPS() {
+    return BaseAPI.socialDomain + 'groups';
+  }
+
+  get SOCIAL_POSTS() {
+    return BaseAPI.socialDomain + 'posts';
+  }
+
+  get SOCIAL_CREATE_POST() {
+    return BaseAPI.socialDomain + 'posts/store';
   }
 }