+++ src/store/Store.js
@@ -271,6 +271,7 @@ class Store {
   @observable app_data = null;
   @observable deep_link_data = null;
   @observable stores_finish = false;
+  @observable selectedFilter = {};
 
   @action setStoresFinish(flag) {
     this.stores_finish = flag;
@@ -296,6 +297,10 @@ class Store {
 
   @action setDeepLinkData(data) {
     this.deep_link_data = data;
+  }
+
+  @action setSelectedFilter(data) {
+    this.selectedFilter = data;
   }
 
   /*********** home begin **********/