+++ CHANGELOG.md
@@ -8,7 +8,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Fixed
-- Disable zoom in webview (notifiItem, Item) by [lamah.dev@gmail.com].
+- Disable zoom in webview (NotifyItem, Item) by [lamah.dev@gmail.com].
+## [r11.6.3] - 2021-05-20
+### Added
+- Update moment locale relative time when change language.
+### Changed
+- Unsubscribe all listener if unmounted in Account.
+- Add site_id for change avatar api.
+### Fixed
+- Not refresh comments when tap bolt.
+- Not correct clip comment if too long in Android.
+- Logic to clip comment if too long.
+- Upload avatar.
 
 ## [r11.6.2] - 2021-05-19
 ### Added